Trait actix_http::body::MessageBody

pub trait MessageBody {
    fn size(&self) -> BodySize;
    fn poll_next(
        self: Pin<&mut Self>,
        cx: &mut Context<'_>
    ) -> Poll<Option<Result<Bytes, Error>>>;
}

Type that provides this trait can be streamed to a peer.

Required Methods

fn size(&self) -> BodySize

fn poll_next(
self: Pin<&mut Self>,
cx: &mut Context<'_>
) -> Poll<Option<Result<Bytes, Error>>>

Implementations

impl dyn MessageBody + 'static

pub fn downcast_ref<T: MessageBody + 'static>(&self) -> Option<&T>

Downcasts generic body to a specific type.

pub fn downcast_mut<T: MessageBody + 'static>(&mut self) -> Option<&mut T>

Downcasts a generic body to a mutable specific type.