Module actix_http::client

Http client api

Structs

Connect

Connector
Manages http client network connectivity The Connector type uses a builder-like combinator pattern for service construction that finishes by calling the .finish() method.

Enums

ConnectError
A set of errors that can occur while connecting to an HTTP host

FreezeRequestError
A set of errors that can occur during freezing a request

InvalidUrl

Protocol
Protocol version

SendRequestError
A set of errors that can occur during request sending and response reading

Traits

Connection