Struct actix_http::client::Connector

pub struct Connector<T> { /* fields omitted */ }

Manages HTTP client network connectivity.

The Connector type uses a builder-like combinator pattern for service construction that finishes by calling the .finish() method.

use std::time::Duration;
use actix_http::client::Connector;

let connector = Connector::new()
     .timeout(Duration::from_secs(5))
     .finish();

Implementations

impl Connector<()>

pub fn new(
) -> Connector<impl Service<TcpConnect<Uri>, Response = TcpConnection<Uri, TcpStream>, Error = ConnectError> + Clone>

impl<S> Connector<S>

pub fn connector<S1, Io1>(self, connector: S1) -> Connector<S1> where
Io1: ActixStream + Debug + 'static,
S1: Service<TcpConnect<Uri>, Response = TcpConnection<Uri, Io1>, Error = TcpConnectError> + Clone,

Use custom connector.

impl<S, Io> Connector<S> where
Io: ActixStream + Debug + 'static,
S: Service<TcpConnect<Uri>, Response = TcpConnection<Uri, Io>, Error = TcpConnectError> + Clone + 'static,

pub fn timeout(self, timeout: Duration) -> Self

Tcp connection timeout, i.e. max time to connect to remote host including dns name resolution. Set to 5 second by default.

pub fn handshake_timeout(self, timeout: Duration) -> Self

Tls handshake timeout, i.e. max time to do tls handshake with remote host after tcp connection established. Set to 5 second by default.

pub fn max_http_version(self, val: Version) -> Self

Maximum supported HTTP major version.

Supported versions are HTTP/1.1 and HTTP/2.

pub fn initial_window_size(self, size: u32) -> Self

Indicates the initial window size (in octets) for HTTP2 stream-level flow control for received data.

The default value is 65,535 and is good for APIs, but not for big objects.

pub fn initial_connection_window_size(self, size: u32) -> Self

Indicates the initial window size (in octets) for HTTP2 connection-level flow control for received data.

The default value is 65,535 and is good for APIs, but not for big objects.

pub fn limit(self, limit: usize) -> Self

Set total number of simultaneous connections per type of scheme.

If limit is 0, the connector has no limit. The default limit size is 100.

pub fn conn_keep_alive(self, dur: Duration) -> Self

Set keep-alive period for opened connection.

Keep-alive period is the period between connection usage. If the delay between repeated usages of the same connection exceeds this period, the connection is closed. Default keep-alive period is 15 seconds.

pub fn conn_lifetime(self, dur: Duration) -> Self

Set max lifetime period for connection.

Connection lifetime is max lifetime of any opened connection until it is closed regardless of keep-alive period. Default lifetime period is 75 seconds.

pub fn disconnect_timeout(self, dur: Duration) -> Self

Set server connection disconnect timeout in milliseconds.

Defines a timeout for disconnect connection. If a disconnect procedure does not complete within this time, the socket get dropped. This timeout affects only secure connections.

To disable timeout set value to 0.

By default disconnect timeout is set to 3000 milliseconds.

pub fn local_address(self, addr: IpAddr) -> Self

Set local IP Address the connector would use for establishing connection.

pub fn finish(self) -> ConnectorService<S, Io>

Finish configuration process and create connector service. The Connector builder always concludes by calling finish() last in its combinator chain.