Trait actix_http::header::IntoHeaderValue

pub trait IntoHeaderValue: Sized {
    type Error: Into<HttpError>;
    fn try_into_value(self) -> Result<HeaderValue, Self::Error>;
}

An interface for types that can be converted into a HeaderValue.

Associated Types

type Error: Into<HttpError>

The type returned in the event of a conversion error.

Required methods

fn try_into_value(self) -> Result<HeaderValue, Self::Error>

Try to convert value to a HeaderValue.