actix_http::header
Trait TryIntoHeaderValue

pub trait TryIntoHeaderValue: Sized {
    type Error: Into<HttpError>;

    // Required method
    fn try_into_value(self) -> Result<HeaderValue, Self::Error>;
}

An interface for types that can be converted into a HeaderValue.

Required Associated Types

type Error: Into<HttpError>

The type returned in the event of a conversion error.

Required Methods

fn try_into_value(self) -> Result<HeaderValue, Self::Error>

Try to convert value to a HeaderValue.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.