Struct actix_http::HttpService

pub struct HttpService<T, P, S, B, X = ExpectHandler, U = UpgradeHandler<T>> { /* fields omitted */ }

NewService HTTP1.1/HTTP2 transport implementation

Methods

impl<T, S, B> HttpService<T, (), S, B> where
S: NewService<Config = SrvConfig, Request = Request>,
S::Error: Into<Error>,
S::InitError: Debug,
S::Response: Into<Response<B>>,
<S::Service as Service>::Future: 'static,
B: MessageBody + 'static,

pub fn build() -> HttpServiceBuilder<T, S>

Create builder for HttpService instance.

impl<T, P, S, B> HttpService<T, P, S, B> where
S: NewService<Config = SrvConfig, Request = Request>,
S::Error: Into<Error>,
S::InitError: Debug,
S::Response: Into<Response<B>>,
<S::Service as Service>::Future: 'static,
B: MessageBody + 'static,

pub fn new<F: IntoNewService<S>>(service: F) -> Self

Create new HttpService instance.

impl<T, P, S, B, X, U> HttpService<T, P, S, B, X, U> where
S: NewService<Config = SrvConfig, Request = Request>,
S::Error: Into<Error>,
S::InitError: Debug,
S::Response: Into<Response<B>>,
B: MessageBody,

pub fn expect<X1>(self, expect: X1) -> HttpService<T, P, S, B, X1, U> where
X1: NewService<Config = SrvConfig, Request = Request, Response = Request>,
X1::Error: Into<Error>,
X1::InitError: Debug,

Provide service for EXPECT: 100-Continue support.

Service get called with request that contains EXPECT header. Service must return request in case of success, in that case request will be forwarded to main service.

pub fn upgrade<U1>(self, upgrade: Option<U1>) -> HttpService<T, P, S, B, X, U1> where
U1: NewService<Config = SrvConfig, Request = (Request, Framed<T, Codec>), Response = ()>,
U1::Error: Display,
U1::InitError: Debug,

Provide service for custom Connection: UPGRADE support.

If service is provided then normal requests handling get halted and this service get called with original request and framed object.