Struct actix_http::Request

pub struct Request<P = BoxedPayloadStream> { /* private fields */ }

An HTTP request.

Implementations

impl Request<BoxedPayloadStream>

pub fn new() -> Request<BoxedPayloadStream>

Create new Request instance

impl<P> Request<P>

pub fn with_payload(payload: Payload<P>) -> Request<P>

Create new Request instance

pub fn replace_payload<P1>( self, payload: Payload<P1> ) -> (Request<P1>, Payload<P>)

Create new Request instance

pub fn payload(&mut self) -> &mut Payload<P>

Get request’s payload

pub fn take_payload(&mut self) -> Payload<P>

Get request’s payload

pub fn into_parts(self) -> (Message<RequestHead>, Payload<P>)

Split request into request head and payload

pub fn head(&self) -> &RequestHead

Http message part of the request

pub fn headers_mut(&mut self) -> &mut HeaderMap

Mutable reference to the message’s headers.

pub fn uri(&self) -> &Uri

Request’s uri.

pub fn uri_mut(&mut self) -> &mut Uri

Mutable reference to the request’s uri.

pub fn method(&self) -> &Method

Read the Request method.

pub fn version(&self) -> Version

Read the Request Version.

pub fn path(&self) -> &str

The target path of this Request.

pub fn upgrade(&self) -> bool

Check if request requires connection upgrade

pub fn peer_addr(&self) -> Option<SocketAddr>

Peer socket address.

Peer address is the directly connected peer’s socket address. If a proxy is used in front of the Actix Web server, then it would be address of this proxy.

Will only return None when called in unit tests unless set manually.

pub fn conn_data<T: 'static>(&self) -> Option<&T>

Returns a reference a piece of connection data set in an on-connect callback.

let opt_t = req.conn_data::<PeerCertificate>();

pub fn take_conn_data(&mut self) -> Option<Rc<Extensions>>

Returns the connection-level data/extensions container if an on-connect callback was registered, leaving an empty one in its place.

pub fn take_req_data(&mut self) -> Extensions

Returns the request-local data/extensions container, leaving an empty one in its place.