Module actix_http::test

Various testing helpers for use in internal and app tests.

Structs

TestBuffer
Async I/O test buffer.

TestRequest
Test Request builder

TestSeqBuffer
Async I/O test buffer with ability to incrementally add to the read buffer.

TestSeqInner