Struct actix_http::uri::Builder

pub struct Builder { /* private fields */ }

A builder for Uris.

This type can be used to construct an instance of Uri through a builder pattern.

Implementations

impl Builder

pub fn new() -> Builder

Creates a new default instance of Builder to construct a Uri.

Examples

let uri = uri::Builder::new()
    .scheme("https")
    .authority("hyper.rs")
    .path_and_query("/")
    .build()
    .unwrap();

pub fn scheme<T>(self, scheme: T) -> Builder
where Scheme: TryFrom<T>, <Scheme as TryFrom<T>>::Error: Into<Error>,

Set the Scheme for this URI.

Examples

let mut builder = uri::Builder::new();
builder.scheme("https");

pub fn authority<T>(self, auth: T) -> Builder
where Authority: TryFrom<T>, <Authority as TryFrom<T>>::Error: Into<Error>,

Set the Authority for this URI.

Examples

let uri = uri::Builder::new()
    .authority("tokio.rs")
    .build()
    .unwrap();

pub fn path_and_query<T>(self, p_and_q: T) -> Builder
where PathAndQuery: TryFrom<T>, <PathAndQuery as TryFrom<T>>::Error: Into<Error>,

Set the PathAndQuery for this URI.

Examples

let uri = uri::Builder::new()
    .path_and_query("/hello?foo=bar")
    .build()
    .unwrap();

pub fn build(self) -> Result<Uri, Error>

Consumes this builder, and tries to construct a valid Uri from the configured pieces.

Errors

This function may return an error if any previously configured argument failed to parse or get converted to the internal representation. For example if an invalid scheme was specified via scheme("!@#%/^") the error will be returned when this function is called rather than when scheme was called.

Additionally, the various forms of URI require certain combinations of parts to be set to be valid. If the parts don’t fit into any of the valid forms of URI, a new error is returned.

Examples

let uri = Uri::builder()
    .build()
    .unwrap();