Crate actix_http

HTTP primitives for the Actix ecosystem.

Crate Features

Feature	Functionality

openssl	TLS support via OpenSSL.

rustls	TLS support via rustls.

compress-brotli	Payload compression support: Brotli.

compress-gzip	Payload compression support: Deflate, Gzip.

compress-zstd	Payload compression support: Zstd.

trust-dns	Use trust-dns as the client DNS resolver.

Re-exports

pub use self::error::Error;

Modules

body
Traits and structures to aid consuming and writing HTTP payloads.

client
HTTP client.

encoding
Content-Encoding support.

error
Error and Result module

h1
HTTP/1 protocol implementation.

h2
HTTP/2 protocol.

header
Pre-defined HeaderNames, traits for parsing and conversion, and other header utility methods.

http
Various HTTP related types.

test
Various testing helpers for use in internal and app tests.

uri
URI component of request and response lines

ws
WebSocket protocol implementation.

Structs

Extensions
A type map for request extensions.

HttpService
A ServiceFactory for HTTP/1.1 or HTTP/2 protocol.

HttpServiceBuilder
A HTTP service builder

Message

Method
The Request Method (VERB)

Request
Request

RequestHead

Response
An HTTP response.

ResponseBuilder
An HTTP response builder.

ResponseHead

ServiceConfig
Http service configuration

StatusCode
An HTTP status code (status-code in RFC 7230 et al.).

Uri
The URI component of a request.

Version
Represents a version of the HTTP spec.

Enums

ConnectionType
Represents various types of connection

ContentEncoding
Represents a supported content encoding.

KeepAlive
Server keep-alive setting

Payload
Type represent streaming payload

Protocol
A major HTTP protocol version.

RequestHeadType

Traits

HttpMessage
Trait that implements general purpose operations on HTTP messages.

Type Definitions

PayloadStream
Type represent boxed payload