Module actix_http::body

Traits and structures to aid consuming and writing HTTP payloads.

“Body” and “payload” are used somewhat interchangeably in this documentation.

Structs

BodyStream

Streaming response wrapper.

BoxBody

A boxed message body with boxed errors.

None

Body type for responses that forbid payloads.

SizedStream

Known sized streaming response wrapper.

Enums

BodySize

Body size hint.

EitherBody

An “either” type specialized for body types.

Traits

MessageBody

An interface for types that can be used as a response body.

Functions

to_bytes

Collects the body produced by a MessageBody implementation into Bytes.