Trait actix_http::body::MessageBody

pub trait MessageBody {
    fn size(&self) -> BodySize;

    fn poll_next(
        self: Pin<&mut Self>, 
        cx: &mut Context<'_>
    ) -> Poll<Option<Result<Bytes, Error>>>;
}

An interface for response bodies.

Required methods

fn size(&self) -> BodySize

Body size hint.

fn poll_next(
self: Pin<&mut Self>,
cx: &mut Context<'_>
) -> Poll<Option<Result<Bytes, Error>>>

Attempt to pull out the next chunk of body bytes.

Loading content...
Implementations

impl dyn MessageBody + 'static

pub fn downcast_ref<T: MessageBody + 'static>(&self) -> Option<&T>

Downcasts generic body to a specific type.

pub fn downcast_mut<T: MessageBody + 'static>(&mut self) -> Option<&mut T>

Downcasts a generic body to a mutable specific type.