Module actix_http::client

HTTP client.

Re-exports

pub use crate::Protocol;

Structs

Connect

Connector
Manages HTTP client network connectivity.

TcpConnect
Connection info.

TcpConnection

Enums

ConnectError
A set of errors that can occur while connecting to an HTTP host

Connection
Unified connection type cover Http1 Plain/Tls and Http2 protocols

FreezeRequestError
A set of errors that can occur during freezing a request

InvalidUrl

SendRequestError
A set of errors that can occur during request sending and response reading

TcpConnectError

Traits

ConnectionIo
Trait alias for types impl tokio::io::AsyncRead and tokio::io::AsyncWrite.

Type Definitions

ConnectorService
Connector service for pooled Plain/Tls Tcp connections.