Module actix_http::error

Error and Result module

Structs

Error

HttpError

A generic “error” for HTTP connections

Enums

ContentTypeError

A set of error that can occur during parsing content type.

DispatchError

A set of errors that can occur during dispatching HTTP requests.

ParseError

A set of errors that can occur during parsing HTTP streams.

PayloadError

A set of errors that can occur during payload parsing.