actix_http

Module error

Error and Result module

Structs

Error

HttpError

A generic “error” for HTTP connections

InvalidStatusCode

A possible error value when converting a StatusCode from a u16 or &str

Enums

ContentTypeError

A set of error that can occur during parsing content type.

DispatchError

A set of errors that can occur during dispatching HTTP requests.

ParseError

A set of errors that can occur during parsing HTTP streams.

PayloadError

A set of errors that can occur during payload parsing.