Module actix_http::error

Error and Result module

Structs

BlockingError
A set of errors that can occur running blocking tasks in thread pool.

Error
General purpose actix web error.

InternalError
Helper type that can wrap any error and generate custom response.

Enums

ContentTypeError
A set of error that can occur during parsing content type

CookieParseError
Enum corresponding to a parsing error.

DispatchError
A set of errors that can occur during dispatching HTTP requests

ParseError
A set of errors that can occur during parsing HTTP streams

PayloadError
A set of errors that can occur during payload parsing

Traits

ResponseError
Error that can be converted to Response

Functions

ErrorBadGateway
Helper function that creates wrapper of any error and generate BAD GATEWAY response.

ErrorBadRequest
Helper function that creates wrapper of any error and generate BAD REQUEST response.

ErrorConflict
Helper function that creates wrapper of any error and generate CONFLICT response.

ErrorExpectationFailed
Helper function that creates wrapper of any error and generate EXPECTATION FAILED response.

ErrorFailedDependency
Helper function that creates wrapper of any error and generate FAILED DEPENDENCY response.

ErrorForbidden
Helper function that creates wrapper of any error and generate FORBIDDEN response.

ErrorGatewayTimeout
Helper function that creates wrapper of any error and generate GATEWAY TIMEOUT response.

ErrorGone
Helper function that creates wrapper of any error and generate GONE response.

ErrorHttpVersionNotSupported
Helper function that creates wrapper of any error and generate HTTP VERSION NOT SUPPORTED response.

ErrorImATeapot
Helper function that creates wrapper of any error and generate IM A TEAPOT response.

ErrorInsufficientStorage
Helper function that creates wrapper of any error and generate INSUFFICIENT STORAGE response.

ErrorInternalServerError
Helper function that creates wrapper of any error and generate INTERNAL SERVER ERROR response.

ErrorLengthRequired
Helper function that creates wrapper of any error and generate LENGTH REQUIRED response.

ErrorLocked
Helper function that creates wrapper of any error and generate LOCKED response.

ErrorLoopDetected
Helper function that creates wrapper of any error and generate LOOP DETECTED response.

ErrorMethodNotAllowed
Helper function that creates wrapper of any error and generate METHOD NOT ALLOWED response.

ErrorMisdirectedRequest
Helper function that creates wrapper of any error and generate MISDIRECTED REQUEST response.

ErrorNetworkAuthenticationRequired
Helper function that creates wrapper of any error and generate NETWORK AUTHENTICATION REQUIRED response.

ErrorNotAcceptable
Helper function that creates wrapper of any error and generate NOT ACCEPTABLE response.

ErrorNotExtended
Helper function that creates wrapper of any error and generate NOT EXTENDED response.

ErrorNotFound
Helper function that creates wrapper of any error and generate NOT FOUND response.

ErrorNotImplemented
Helper function that creates wrapper of any error and generate NOT IMPLEMENTED response.

ErrorPayloadTooLarge
Helper function that creates wrapper of any error and generate PAYLOAD TOO LARGE response.

ErrorPaymentRequired
Helper function that creates wrapper of any error and generate PAYMENT_REQUIRED response.

ErrorPreconditionFailed
Helper function that creates wrapper of any error and generate PRECONDITION FAILED response.

ErrorPreconditionRequired
Helper function that creates wrapper of any error and generate PRECONDITION REQUIRED response.

ErrorProxyAuthenticationRequired
Helper function that creates wrapper of any error and generate PROXY AUTHENTICATION REQUIRED response.

ErrorRangeNotSatisfiable
Helper function that creates wrapper of any error and generate RANGE NOT SATISFIABLE response.

ErrorRequestHeaderFieldsTooLarge
Helper function that creates wrapper of any error and generate REQUEST HEADER FIELDS TOO LARGE response.

ErrorRequestTimeout
Helper function that creates wrapper of any error and generate REQUEST TIMEOUT response.

ErrorServiceUnavailable
Helper function that creates wrapper of any error and generate SERVICE UNAVAILABLE response.

ErrorTooManyRequests
Helper function that creates wrapper of any error and generate TOO MANY REQUESTS response.

ErrorUnauthorized
Helper function that creates wrapper of any error and generate UNAUTHORIZED response.

ErrorUnavailableForLegalReasons
Helper function that creates wrapper of any error and generate UNAVAILABLE FOR LEGAL REASONS response.

ErrorUnprocessableEntity
Helper function that creates wrapper of any error and generate UNPROCESSABLE ENTITY response.

ErrorUnsupportedMediaType
Helper function that creates wrapper of any error and generate UNSUPPORTED MEDIA TYPE response.

ErrorUpgradeRequired
Helper function that creates wrapper of any error and generate UPGRADE REQUIRED response.

ErrorUriTooLong
Helper function that creates wrapper of any error and generate URI TOO LONG response.

ErrorVariantAlsoNegotiates
Helper function that creates wrapper of any error and generate VARIANT ALSO NEGOTIATES response.

Type Definitions

Result
A specialized std::result::Result for actix web operations