Module actix_http::error

Error and Result module

Structs

BlockingError
A set of errors that can occur running blocking tasks in thread pool.

Error

HttpError
A generic “error” for HTTP connections

Enums

ContentTypeError
A set of error that can occur during parsing content type.

DispatchError
A set of errors that can occur during dispatching HTTP requests.

Kind

ParseError
A set of errors that can occur during parsing HTTP streams.

PayloadError
A set of errors that can occur during payload parsing.