Trait actix_http::header::TryIntoHeaderValue

pub trait TryIntoHeaderValue: Sized {
    type Error: Into<HttpError>;

    fn try_into_value(self) -> Result<HeaderValue, Self::Error>;
}

An interface for types that can be converted into a HeaderValue.

Required Associated Types

type Error: Into<HttpError>

The type returned in the event of a conversion error.

Required Methods

fn try_into_value(self) -> Result<HeaderValue, Self::Error>

Try to convert value to a HeaderValue.