Struct actix_http::Request

pub struct Request<P = PayloadStream> { /* fields omitted */ }

Request

Implementations

impl Request<PayloadStream>

pub fn new() -> Request<PayloadStream>

Create new Request instance

impl<P> Request<P>

pub fn with_payload(payload: Payload<P>) -> Request<P>

Create new Request instance

pub fn replace_payload<P1>(
self,
payload: Payload<P1>
) -> (Request<P1>, Payload<P>)

Create new Request instance

pub fn payload(&mut self) -> &mut Payload<P>

Get request’s payload

pub fn take_payload(&mut self) -> Payload<P>

Get request’s payload

pub fn into_parts(self) -> (Message<RequestHead>, Payload<P>)

Split request into request head and payload

pub fn head(&self) -> &RequestHead

Http message part of the request

pub fn headers_mut(&mut self) -> &mut HeaderMap

Mutable reference to the message’s headers.

pub fn uri(&self) -> &Uri

Request’s uri.

pub fn uri_mut(&mut self) -> &mut Uri

Mutable reference to the request’s uri.

pub fn method(&self) -> &Method

Read the Request method.

pub fn version(&self) -> Version

Read the Request Version.

pub fn path(&self) -> &str

The target path of this Request.

pub fn upgrade(&self) -> bool

Check if request requires connection upgrade

pub fn peer_addr(&self) -> Option<SocketAddr>

Peer socket address.

Peer address is the directly connected peer’s socket address. If a proxy is used in front of the Actix Web server, then it would be address of this proxy.

Will only return None when called in unit tests.