Trait actix_http::HttpMessage

pub trait HttpMessage: Sized {
    type Stream;

    fn headers(&self) -> &HeaderMap;
    fn take_payload(&mut self) -> Payload<Self::Stream>;
    fn extensions(&self) -> Ref<'_, Extensions>;
    fn extensions_mut(&self) -> RefMut<'_, Extensions>;

    fn content_type(&self) -> &str { ... }
    fn encoding(&self) -> Result<&'static Encoding, ContentTypeError> { ... }
    fn mime_type(&self) -> Result<Option<Mime>, ContentTypeError> { ... }
    fn chunked(&self) -> Result<bool, ParseError> { ... }
}

Trait that implements general purpose operations on HTTP messages.

Required Associated Types

type Stream

Type of message payload stream

Required Methods

fn headers(&self) -> &HeaderMap

Read the message headers.

fn take_payload(&mut self) -> Payload<Self::Stream>

Message payload stream

fn extensions(&self) -> Ref<'_, Extensions>

Returns a reference to the request-local data/extensions container.

fn extensions_mut(&self) -> RefMut<'_, Extensions>

Returns a mutable reference to the request-local data/extensions container.

Provided Methods

fn content_type(&self) -> &str

Read the request content type. If request did not contain a Content-Type header, an empty string is returned.

fn encoding(&self) -> Result<&'static Encoding, ContentTypeError>

Get content type encoding.

UTF-8 is used by default, If request charset is not set.

fn mime_type(&self) -> Result<Option<Mime>, ContentTypeError>

Convert the request content type to a known mime type.

fn chunked(&self) -> Result<bool, ParseError>

Check if request has chunked transfer encoding.