Module actix_http::ws

WebSocket protocol implementation.

To setup a WebSocket, first perform the WebSocket handshake then on success convert Payload into a WsStream stream and then use WsWriter to communicate with the peer.

Structs

CloseReason
Reason for closing the connection

Codec
WebSocket protocol codec.

Dispatcher

Parser
A struct representing a WebSocket frame.

Enums

CloseCode
Status code used to indicate why an endpoint is closing the WebSocket connection.

Frame
A WebSocket frame.

HandshakeError
WebSocket handshake errors

Item
A WebSocket continuation item.

Message
A WebSocket message.

OpCode
Operation codes as part of RFC6455.

ProtocolError
WebSocket protocol errors.

Functions

handshake
Verify WebSocket handshake request and create handshake response.

handshake_response
Create WebSocket handshake response.

hash_key
Hashes the Sec-WebSocket-Key header according to the WebSocket spec.

verify_handshake
Verify WebSocket handshake request.